Kludges to the rest of the compiler following the change to the parser.
I had made a change (da5550f) to the parser to introduce the parsing
of meta-brackets (".<" ">."). This commit contains (completely
misguided) changes I had to make to the rest of the compiler to
"hammer through" the changes to the parser.  These changes get rid of
compile-time warnings (which are fatal in the default build settings,
and I'd like to keep it that way), but the effect they have on the
rest of the compiler is not "The Right Thing To Do."