Avoid a crash when the toploop is used to run a script.
Register the OCaml side of the Run primitive when the toploop is used
to execute a script (i.e., using the "-stdin" option).  Before I did
this, the C side of the Run primitive would crash because it was
trying to dereferenc a null pointer (because Callback.register wasn't
getting called to register the OCaml side of the Run primitive).