"Make the native standalone executables work."
The problem was that, when Opttoploop.load_lambda called dll_run
(which, in turn called caml_natdynlink_run_toplevel, which called
caml_natdynlink_run), load_lambda was passing the string "TOP" to
caml_natdynlink_run() as the name of the compilation unit (the second
argument to caml_natdynlink_run(), i.e., "value symbol").  On the
other hand, when load_lambda was compiling the shared library that was
going to be passed to dll_run, the name of the compilation unit was
the empty string (because the current value of
Compilenv.current_unit_name() was the empty string).

Running from inside opttopnat didn't have that problem because
Opttoploop.execute_phrase calls Compilenv.reset.

For now, I am just calling Compilenv.reset from metaocaml.ml.