Package the closure and the code into a single block.
When compiling values of type "code" the closure that "captures" the
variables which occur free in the  body of the lambda-term that is "the
code" and the actual code are now packaged into a single heap block.
(Before, they were returned as two separate values on the byte code
interpreter's stack: returning them as two separate values not only made
little sense from "the point of view of semantics", but it also caused
crashes when even simple expressions were evaluated in toplevel).